requirements.txt
numpy
pandas
opencv-python
scikit-image
scikit-learn
xgboost
shap
matplotlib
tqdm
pillow
streamlit
joblib
